+++ script/compiled/dev/commands.js
@@ -2,15 +2,15 @@
 const developerCommands = [
     {
         name: "help",
-        description: "List all available commands",
-        help: "help [command] - List all available commands. If a command is specified, the help text for that command will be displayed.",
-        list: [],
+        help: "List all available commands",
+        verboseHelp: "help [command] - List all available commands. If a command is specified, the help text for that command will be displayed.",
+        availableParams: [],
         execute: (args) => {
             const command = args[0];
             if (command) {
                 const commandObject = developerCommands.find((c) => c.name === command);
                 if (commandObject) {
-                    devConsole.commandHistory.push(commandObject.help);
+                    devConsole.commandHistory.push(commandObject.verboseHelp);
                     return;
                 }
                 devConsole.commandHistory.push(`Command ${command} not found, type "help" to see all available commands`);
@@ -19,7 +19,7 @@ const developerCommands = [
             else {
                 let text = "Available commands:";
                 developerCommands.forEach((command) => {
-                    text += `<br>${command.name}: ${command.description}`;
+                    text += `<br>${command.name}: ${command.help}`;
                 });
                 devConsole.commandHistory.push(text);
             }
@@ -27,9 +27,9 @@ const developerCommands = [
     },
     {
         name: "clear",
-        description: "Clear the console",
-        help: "Clears the console, while keeping the command history.",
-        list: [],
+        help: "Clear the console",
+        verboseHelp: "Clears the console, while keeping the command history.",
+        availableParams: [],
         execute: () => {
             devConsole.commandHistory = [""]; // This only clears the logs, not the actual command history
             consoleLog.innerHTML = "";
@@ -37,9 +37,9 @@ const developerCommands = [
     },
     {
         name: "dev",
-        description: "Enable developer mode",
-        help: "dev [save] - Enable developer mode. If 'save' is specified, the setting will be saved to local storage.<br>Use 'dev' again to disable developer mode.<br>Developer mode enables cheat commands and removes certain limitations.",
-        list: [[{ id: "save" }]],
+        help: "Enable developer mode",
+        verboseHelp: "dev [save] - Enable developer mode. If 'save' is specified, the setting will be saved to local storage.<br>Use 'dev' again to disable developer mode.<br>Developer mode enables cheat commands and removes certain limitations.",
+        availableParams: [[{ id: "save" }]],
         execute: (args) => {
             DEVTOOLS.ENABLED = !DEVTOOLS.ENABLED;
             const saveToLocalStorage = args[0] === "save";
@@ -54,10 +54,10 @@ const developerCommands = [
     },
     {
         name: "onepunch",
-        description: "Deal massive damage with every attack",
-        help: "onepunch - Deal massive damage with every attack.<br>Use 'onepunch' again to disable this cheat.",
+        help: "Deal massive damage with every attack",
+        verboseHelp: "onepunch - Deal massive damage with every attack.<br>Use 'onepunch' again to disable this cheat.",
         isCheat: true,
-        list: [],
+        availableParams: [],
         execute: () => {
             DEVTOOLS.ONE_PUNCH = !DEVTOOLS.ONE_PUNCH;
             devConsole.commandHistory.push(`${DEVTOOLS.ONE_PUNCH ? "You have become ONE" : "Your power has been taken away"}`);
@@ -65,10 +65,10 @@ const developerCommands = [
     },
     {
         name: "nocd",
-        description: "Remove cooldowns from all abilities",
-        help: "nocd - Remove cooldowns from all abilities.<br>Use 'nocd' again to disable this cheat.",
+        help: "Remove cooldowns from all abilities",
+        verboseHelp: "nocd - Remove cooldowns from all abilities.<br>Use 'nocd' again to disable this cheat.",
         isCheat: true,
-        list: [],
+        availableParams: [],
         execute: () => {
             DEVTOOLS.NO_CD = !DEVTOOLS.NO_CD;
             devConsole.commandHistory.push(`Skills are now ${DEVTOOLS.NO_CD ? "instant" : "lethargic"}`);
@@ -76,10 +76,10 @@ const developerCommands = [
     },
     {
         name: "freecast",
-        description: "Cast abilities without spending mana",
-        help: "freecast - Cast abilities without spending mana.<br>Use 'freecast' again to disable this cheat.",
+        help: "Cast abilities without spending mana",
+        verboseHelp: "freecast - Cast abilities without spending mana.<br>Use 'freecast' again to disable this cheat.",
         isCheat: true,
-        list: [],
+        availableParams: [],
         execute: () => {
             DEVTOOLS.FREE_CAST = !DEVTOOLS.FREE_CAST;
             devConsole.commandHistory.push(`Your mana is now ${DEVTOOLS.FREE_CAST ? "infinite" : "very much finite"}`);
@@ -87,10 +87,10 @@ const developerCommands = [
     },
     {
         name: "god",
-        description: "Become invincible",
-        help: "god - All damage dealt to you will be 0, and you are immune to debuffs.<br>Use 'god' again to disable this cheat.",
+        help: "Become invincible",
+        verboseHelp: "god - All damage dealt to you will be 0, and you are immune to debuffs.<br>Use 'god' again to disable this cheat.",
         isCheat: true,
-        list: [],
+        availableParams: [],
         execute: () => {
             DEVTOOLS.GOD = !DEVTOOLS.GOD;
             devConsole.commandHistory.push(`Invincibility ${DEVTOOLS.GOD ? "granted" : "taken away"}`);
@@ -98,10 +98,10 @@ const developerCommands = [
     },
     {
         name: "ignore",
-        description: "[category] Ignore specific requirements",
-        help: "ignore [category] - Ignore certain requirements in the game. Arguments:<br>leveling - ignores all requirements for perks/class/skills<br>Example: ignore leveling.",
+        help: "[category] Ignore specific requirements",
+        verboseHelp: "ignore [category] - Ignore certain requirements in the game. Arguments:<br>leveling - ignores all requirements for perks/class/skills<br>Example: ignore leveling.",
         isCheat: true,
-        list: [[{ id: "leveling" }]],
+        availableParams: [[{ id: "leveling" }]],
         execute: (args) => {
             const cat = args[0];
             if (cat) {
@@ -120,10 +120,10 @@ const developerCommands = [
     },
     {
         name: "enter",
-        description: "[dungeon] Enter a dungeon",
-        help: "enter [dungeon] - Enter a specified dungeon.<br>Example: enter vithail_dungeon",
+        help: "[dungeon] Enter a dungeon",
+        verboseHelp: "enter [dungeon] - Enter a specified dungeon.<br>Example: enter vithail_dungeon",
         isCheat: true,
-        list: [dungeons],
+        availableParams: [dungeons],
         execute: (args) => {
             const dungeon = args[0];
             if (dungeon) {
@@ -143,10 +143,10 @@ const developerCommands = [
     },
     {
         name: "leave",
-        description: "Leave the current dungeon",
-        help: "leave - Leave the current dungeon.",
+        help: "Leave the current dungeon",
+        verboseHelp: "leave - Leave the current dungeon.",
         isCheat: true,
-        list: [],
+        availableParams: [],
         execute: () => {
             if (dungeonController.currentDungeon) {
                 devConsole.commandHistory.push(`Left dungeon ${dungeonController.currentDungeon.id}`);
@@ -159,10 +159,10 @@ const developerCommands = [
     },
     {
         name: "fight",
-        description: "[enemy] [amount] Fight an enemy",
-        help: "fight [enemy] [amount] - Fight an enemy. Specify amount for group battle.<br>Example: fight skeleton 2",
+        help: "[enemy] [amount] Fight an enemy",
+        verboseHelp: "fight [enemy] [amount] - Fight an enemy. Specify amount for group battle.<br>Example: fight skeleton 2",
         isCheat: true,
-        list: [Object.values(enemies), [{ id: "amount - number", onSelect: "1" }]],
+        availableParams: [Object.values(enemies), [{ id: "amount - number", onSelect: "1" }]],
         execute: (args) => {
             const enemy = args[0];
             const amount = args[1] ? parseInt(args[1]) : 1;
@@ -185,8 +185,8 @@ const developerCommands = [
     },
     {
         name: "lose",
-        description: "Lose the current fight",
-        help: "lose - Reduces your hp to 0, causing you to be defeated.",
+        help: "Lose the current fight",
+        verboseHelp: "lose - Reduces your hp to 0, causing you to be defeated.",
         execute: () => {
             player.stats.hp = 0;
             update();
@@ -195,10 +195,10 @@ const developerCommands = [
     },
     {
         name: "killall",
-        description: "Kill all enemies in the current fight.",
-        help: "killall - Kill all enemies in the current fight.",
+        help: "Kill all enemies in the current fight.",
+        verboseHelp: "killall - Kill all enemies in the current fight.",
         isCheat: true,
-        list: [],
+        availableParams: [],
         execute: () => {
             const _enemies = combat.getLivingEnemies();
             if (_enemies.length > 0) {
@@ -216,10 +216,10 @@ const developerCommands = [
     },
     {
         name: "item",
-        description: "[item] [amount] Adds specified quantity of items to your inventory",
-        help: "item [item] [amount] - Adds specified quantity of items to your inventory.<br>Example: item small_healing_potion 3",
+        help: "[item] [amount] Adds specified quantity of items to your inventory",
+        verboseHelp: "item [item] [amount] - Adds specified quantity of items to your inventory.<br>Example: item small_healing_potion 3",
         isCheat: true,
-        list: [Object.values(items), [{ id: "amount - number", onSelect: "5" }]],
+        availableParams: [Object.values(items), [{ id: "amount - number", onSelect: "5" }]],
         execute: (args) => {
             const item = args[0];
             const quantity = args[1] ? parseInt(args[1]) : 1;
@@ -243,10 +243,10 @@ const developerCommands = [
     },
     {
         name: "gold",
-        description: "Add gold",
-        help: "gold [amount] - Add gold to your inventory.<br>Example: gold 100",
+        help: "Add gold",
+        verboseHelp: "gold [amount] - Add gold to your inventory.<br>Example: gold 100",
         isCheat: true,
-        list: [[{ id: "amount - number", onSelect: "100" }]],
+        availableParams: [[{ id: "amount - number", onSelect: "100" }]],
         execute: (args) => {
             const amount = args[0] ? parseInt(args[0]) : 100;
             player.addGold(amount);
@@ -256,10 +256,10 @@ const developerCommands = [
     },
     {
         name: "xp",
-        description: "Add experience points",
-        help: "xp [amount] - Add experience points.<br>Example: xp 100",
+        help: "Add experience points",
+        verboseHelp: "xp [amount] - Add experience points.<br>Example: xp 100",
         isCheat: true,
-        list: [[{ id: "amount - number", onSelect: "50" }]],
+        availableParams: [[{ id: "amount - number", onSelect: "50" }]],
         execute: (args) => {
             const xp = args[0] ? parseInt(args[0]) : 0;
             player.addXP(xp);
@@ -269,10 +269,10 @@ const developerCommands = [
     },
     {
         name: "sp",
-        description: "Add skill points",
-        help: "sp [amount] - Add skill points.<br>Example: sp 5",
+        help: "Add skill points",
+        verboseHelp: "sp [amount] - Add skill points.<br>Example: sp 5",
         isCheat: true,
-        list: [[{ id: "amount - number", onSelect: "5" }]],
+        availableParams: [[{ id: "amount - number", onSelect: "5" }]],
         execute: (args) => {
             const sp = args[0] ? parseInt(args[0]) : 0;
             player.skill_points += sp;
@@ -282,10 +282,10 @@ const developerCommands = [
     },
     {
         name: "pp",
-        description: "Add perk points",
-        help: "pp [amount] - Add perk points.<br>Example: pp 5",
+        help: "Add perk points",
+        verboseHelp: "pp [amount] - Add perk points.<br>Example: pp 5",
         isCheat: true,
-        list: [[{ id: "amount - number", onSelect: "5" }]],
+        availableParams: [[{ id: "amount - number", onSelect: "5" }]],
         execute: (args) => {
             const sp = args[0] ? parseInt(args[0]) : 0;
             player.perk_points += sp;