+++ script/main/icons.ts
@@ -28,6 +28,8 @@ const icons: any = {
   power: "gfx/status/power.png",
   speed: "gfx/status/speedometer.png",
   dodge: "gfx/status/dodge_icon.png",
+  accuracy: "gfx/status/bullseye.png",
+  acc: "gfx/status/bullseye.png",
   critRate: "gfx/status/power.png",
   critPower: "gfx/status/power.png",
   weapon: "gfx/icons/piercing-sword.png",
@@ -54,4 +56,7 @@ coreCharacterStats.forEach((stat) => {
   icons[`hpMaxFrom${stat}`] = `gfx/status/heart-plus.png`;
   icons[`mpMaxFrom${stat}`] = `gfx/icons/magic-swirl.png`;
   icons[`mpRegenFrom${stat}`] = `gfx/icons/regeneration.png`;
+  icons[`accFrom${stat}`] = `gfx/status/bullseye.png`;
+  icons[`critRateFrom${stat}`] = `gfx/status/power.png`;
+  icons[`critPowerFrom${stat}`] = `gfx/status/power.png`;
 });