+++ script/main/save.ts
@@ -8,7 +8,7 @@ class SaveController {
   }
 
   sortSaves() {
-    this.saveSlots.sort((a: SaveFile, b: SaveFile) => b.lastSaved.getTime() - a.lastSaved.getTime());
+    this.saveSlots = this.saveSlots.sort((a: SaveFile, b: SaveFile) => b.lastSaved.getTime() - a.lastSaved.getTime());
   }
 
   getSaves(options?: { update?: boolean }) {
@@ -41,6 +41,7 @@ class SaveController {
     } else {
       this.saveSlots.push(saveFile);
     }
+    this.sortSaves();
     localStorage.setItem("PohjolaIII_saved_games", JSON.stringify(this.saveSlots));
     if (!options?.auto) {
       closeConfirmationWindow();
@@ -84,8 +85,19 @@ class SaveController {
           createLobby();
           createCharView();
           sideBarDetails();
+          game.playing = true;
           log.createNotification(`Loaded save file: ${save.name}`, -1);
           log.write(`<c>white<c>Loaded save file: <c>goldenrod<c>${save.name}`);
+
+          // See if there are any updates that need to be applied to the save file
+          if (parseFloat(save.version) < 0.21) {
+            log.write(`<c>white<c>Updating save file to version <c>lime<c>${gameVersion}`);
+            Object.entries(challenges).forEach(([key, value]: [string, number | boolean]) => {
+              if (typeof value === "number") {
+                challenges[key] += 3;
+              }
+            });
+          }
         } catch (err) {
           log.createNotification("Could not load save file.", -1);
           log.write(`<c>red<c>Error loading save file: ${err}`);
@@ -258,6 +270,7 @@ function saveScreen(menu: boolean = false) {
       "save_to_file"
     )}</button>
         <button class="save-button" onClick="saveController.loadFromFile()">${game.getLocalizedString("load_from_file")}</button>
+        <button class="save-button" onClick="gotoMainMenu()">${game.getLocalizedString("main_menu")}</button>
       </div>
       <div class="save-list"></div>
     `;